+++ src/pos/lmn.rs
@@ -18,7 +18,7 @@ use super::uvw::UVW;
 /// This coordinate system is discussed at length in Interferometry and
 /// Synthesis in Radio Astronomy, Third Edition, Section 3: Analysis of the
 /// Interferometer Response.
-#[derive(Clone, Copy, Debug, Default)]
+#[derive(Clone, Copy, Debug, Default, PartialEq)]
 #[allow(clippy::upper_case_acronyms)]
 pub struct LMN {
     /// l coordinate \[dimensionless\]
@@ -49,10 +49,50 @@ impl LMN {
     }
 }
 
+#[cfg(any(test, feature = "approx"))]
+impl approx::AbsDiffEq for LMN {
+    type Epsilon = f64;
+
+    fn default_epsilon() -> f64 {
+        f64::EPSILON
+    }
+
+    fn abs_diff_eq(&self, other: &Self, epsilon: f64) -> bool {
+        f64::abs_diff_eq(&self.l, &other.l, epsilon)
+            && f64::abs_diff_eq(&self.m, &other.m, epsilon)
+            && f64::abs_diff_eq(&self.n, &other.n, epsilon)
+    }
+}
+
+#[cfg(any(test, feature = "approx"))]
+impl approx::RelativeEq for LMN {
+    #[inline]
+    fn default_max_relative() -> f64 {
+        f64::EPSILON
+    }
+
+    #[inline]
+    fn relative_eq(&self, other: &Self, epsilon: f64, max_relative: f64) -> bool {
+        f64::relative_eq(&self.l, &other.l, epsilon, max_relative)
+            && f64::relative_eq(&self.m, &other.m, epsilon, max_relative)
+            && f64::relative_eq(&self.n, &other.n, epsilon, max_relative)
+    }
+
+    #[inline]
+    fn relative_ne(
+        &self,
+        other: &Self,
+        epsilon: Self::Epsilon,
+        max_relative: Self::Epsilon,
+    ) -> bool {
+        !Self::relative_eq(self, other, epsilon, max_relative)
+    }
+}
+
 /// A "radio interferometer measurement equation (RIME)"-ready version of
 /// [`LMN`]; i.e. `LmnRime.l == 2 * pi * LMN.l`, `LmnRime.m == 2 * pi * LMN.m`,
 /// `LmnRime.n == 2 * pi * (LMN.n - 1)`.
-#[derive(Clone, Copy, Debug, Default)]
+#[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub struct LmnRime {
     /// 2 * pi * l \[dimensionless\]
     pub l: f64,
@@ -76,6 +116,46 @@ impl LmnRime {
             m: self.m / TAU,
             n: self.n / TAU + 1.0,
         }
+    }
+}
+
+#[cfg(any(test, feature = "approx"))]
+impl approx::AbsDiffEq for LmnRime {
+    type Epsilon = f64;
+
+    fn default_epsilon() -> f64 {
+        f64::EPSILON
+    }
+
+    fn abs_diff_eq(&self, other: &Self, epsilon: f64) -> bool {
+        f64::abs_diff_eq(&self.l, &other.l, epsilon)
+            && f64::abs_diff_eq(&self.m, &other.m, epsilon)
+            && f64::abs_diff_eq(&self.n, &other.n, epsilon)
+    }
+}
+
+#[cfg(any(test, feature = "approx"))]
+impl approx::RelativeEq for LmnRime {
+    #[inline]
+    fn default_max_relative() -> f64 {
+        f64::EPSILON
+    }
+
+    #[inline]
+    fn relative_eq(&self, other: &Self, epsilon: f64, max_relative: f64) -> bool {
+        f64::relative_eq(&self.l, &other.l, epsilon, max_relative)
+            && f64::relative_eq(&self.m, &other.m, epsilon, max_relative)
+            && f64::relative_eq(&self.n, &other.n, epsilon, max_relative)
+    }
+
+    #[inline]
+    fn relative_ne(
+        &self,
+        other: &Self,
+        epsilon: Self::Epsilon,
+        max_relative: Self::Epsilon,
+    ) -> bool {
+        !Self::relative_eq(self, other, epsilon, max_relative)
     }
 }
 
@@ -108,9 +188,14 @@ mod tests {
             n: 0.707,
         };
         let lmn_rime = lmn.prepare_for_rime();
-        assert_abs_diff_eq!(lmn_rime.l, PI);
-        assert_abs_diff_eq!(lmn_rime.m, PI);
-        assert_abs_diff_eq!(lmn_rime.n, -1.840973295003619);
+        assert_abs_diff_eq!(
+            lmn_rime,
+            LmnRime {
+                l: PI,
+                m: PI,
+                n: -1.840973295003619
+            }
+        );
     }
 
     #[test]
@@ -137,8 +222,6 @@ mod tests {
         };
         let lmn_rime = lmn.prepare_for_rime();
         let lmn2 = lmn_rime.to_lmn();
-        assert_abs_diff_eq!(lmn.l, lmn2.l);
-        assert_abs_diff_eq!(lmn.m, lmn2.m);
-        assert_abs_diff_eq!(lmn.n, lmn2.n);
+        assert_abs_diff_eq!(lmn, lmn2);
     }
 }