+++ RELEASES.md
@@ -1,5 +1,10 @@
 <!-- markdownlint-disable=MD025 -->
 
+# Unreleased
+
+- Add an optional `approx` feature that exposes trait implementations like
+  `approx::AbsDiffEq` on each of the coordinate types (e.g. `UVW`).
+- Make `Jones` `#[repr(transparent)]`
 # Version 0.7.2 (2022-08-04)
 
 - Expose Marlu version in `built_info`